AutoTutor/requirements.txt
requests
beautifulsoup4
googlesearch-python
regex
nltk
tiktoken
configparser
openai
pinecone-sdk
urllib3
wikipedia-api
google-api-python-client
python-docx
openpyxl
pandas
chardet
pdfminer.six
tqdm
